+++ packages/assets-controllers/src/AccountTrackerController.ts
@@ -8,14 +8,14 @@ import type {
 import type {
   ControllerStateChangeEvent,
   ControllerGetStateAction,
-  RestrictedMessenger,
-} from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
 import {
   query,
   safelyExecuteWithTimeout,
   toChecksumHexAddress,
 } from '@metamask/controller-utils';
 import EthQuery from '@metamask/eth-query';
+import type { Messenger } from '@metamask/messenger';
 import type {
   NetworkClient,
   NetworkClientId,
@@ -196,12 +196,10 @@ export type AllowedEvents =
 /**
  * The messenger of the {@link AccountTrackerController}.
  */
-export type AccountTrackerControllerMessenger = RestrictedMessenger<
+export type AccountTrackerControllerMessenger = Messenger<
   typeof controllerName,
   AccountTrackerControllerActions | AllowedActions,
-  AccountTrackerControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  AccountTrackerControllerEvents | AllowedEvents
 >;
 
 /** The input to start polling for the {@link AccountTrackerController} */
@@ -234,7 +232,7 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
    * @param options - The controller options.
    * @param options.interval - Polling interval used to fetch new account balances.
    * @param options.state - Initial state to set on this controller.
-   * @param options.messenger - The controller messaging system.
+   * @param options.messenger - The controller messenger.
    * @param options.getStakedBalanceForChain - The function to get the staked native asset balance for a chain.
    * @param options.includeStakedAssets - Whether to include staked assets in the account balances.
    * @param options.accountsApiChainIds - Array of chainIds that should use Accounts-API strategy (if supported by API).
@@ -296,7 +294,7 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
 
     this.setIntervalLength(interval);
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'AccountsController:selectedEvmAccountChange',
       (newAddress, prevAddress) => {
         if (newAddress !== prevAddress) {
@@ -313,12 +311,12 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
 
   private syncAccounts(newChainIds: string[]) {
     const accountsByChainId = cloneDeep(this.state.accountsByChainId);
-    const { selectedNetworkClientId } = this.messagingSystem.call(
+    const { selectedNetworkClientId } = this.messenger.call(
       'NetworkController:getState',
     );
     const {
       configuration: { chainId: currentChainId },
-    } = this.messagingSystem.call(
+    } = this.messenger.call(
       'NetworkController:getNetworkClientById',
       selectedNetworkClientId,
     );
@@ -338,7 +336,7 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
     // Note: The address from the preferences controller are checksummed
     // The addresses from the accounts controller are lowercased
     const addresses = Object.values(
-      this.messagingSystem
+      this.messenger
         .call('AccountsController:listAccounts')
         .map((internalAccount) =>
           toChecksumHexAddress(internalAccount.address),
@@ -372,12 +370,12 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
   }
 
   readonly #getProvider = (chainId: Hex): Web3Provider => {
-    const { networkConfigurationsByChainId } = this.messagingSystem.call(
+    const { networkConfigurationsByChainId } = this.messenger.call(
       'NetworkController:getState',
     );
     const cfg = networkConfigurationsByChainId[chainId];
     const { networkClientId } = cfg.rpcEndpoints[cfg.defaultRpcEndpointIndex];
-    const client = this.messagingSystem.call(
+    const client = this.messenger.call(
       'NetworkController:getNetworkClientById',
       networkClientId,
     );
@@ -385,12 +383,12 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
   };
 
   readonly #getNetworkClient = (chainId: Hex) => {
-    const { networkConfigurationsByChainId } = this.messagingSystem.call(
+    const { networkConfigurationsByChainId } = this.messenger.call(
       'NetworkController:getState',
     );
     const cfg = networkConfigurationsByChainId[chainId];
     const { networkClientId } = cfg.rpcEndpoints[cfg.defaultRpcEndpointIndex];
-    return this.messagingSystem.call(
+    return this.messenger.call(
       'NetworkController:getNetworkClientById',
       networkClientId,
     );
@@ -431,13 +429,12 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
   #getCorrectNetworkClient(networkClientId?: NetworkClientId) {
     const selectedNetworkClientId =
       networkClientId ??
-      this.messagingSystem.call('NetworkController:getState')
-        .selectedNetworkClientId;
+      this.messenger.call('NetworkController:getState').selectedNetworkClientId;
     const {
       configuration: { chainId },
       provider,
       blockTracker,
-    } = this.messagingSystem.call(
+    } = this.messenger.call(
       'NetworkController:getNetworkClientById',
       selectedNetworkClientId,
     );
@@ -456,7 +453,7 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
    * @returns An array of network client IDs.
    */
   #getNetworkClientIds(): NetworkClientId[] {
-    const { networkConfigurationsByChainId } = this.messagingSystem.call(
+    const { networkConfigurationsByChainId } = this.messenger.call(
       'NetworkController:getState',
     );
     return Object.values(networkConfigurationsByChainId).flatMap(
@@ -495,13 +492,11 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
     networkClientIds: NetworkClientId[],
     queryAllAccounts: boolean = false,
   ) {
-    const selectedAccount = this.messagingSystem.call(
+    const selectedAccount = this.messenger.call(
       'AccountsController:getSelectedAccount',
     );
-    const allAccounts = this.messagingSystem.call(
-      'AccountsController:listAccounts',
-    );
-    const { isMultiAccountBalancesEnabled } = this.messagingSystem.call(
+    const allAccounts = this.messenger.call('AccountsController:listAccounts');
+    const { isMultiAccountBalancesEnabled } = this.messenger.call(
       'PreferencesController:getState',
     );
 
@@ -762,12 +757,12 @@ export class AccountTrackerController extends StaticIntervalPollingController<Ac
   }
 
   #registerMessageHandlers() {
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:updateNativeBalances` as const,
       this.updateNativeBalances.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:updateStakedBalances` as const,
       this.updateStakedBalances.bind(this),
     );