+++ packages/bridge-controller/src/types.ts
@@ -5,9 +5,10 @@ import type {
   TokenRatesControllerGetStateAction,
 } from '@metamask/assets-controllers';
 import type {
+  ControllerGetStateAction,
   ControllerStateChangeEvent,
-  RestrictedMessenger,
-} from '@metamask/base-controller';
+} from '@metamask/base-controller/next';
+import type { Messenger } from '@metamask/messenger';
 import type {
   NetworkControllerFindNetworkClientIdByChainIdAction,
   NetworkControllerGetStateAction,
@@ -326,8 +327,19 @@ export type BridgeControllerAction<
   handler: BridgeController[FunctionName];
 };
 
+export type BridgeControllerGetStateAction = ControllerGetStateAction<
+  typeof BRIDGE_CONTROLLER_NAME,
+  BridgeControllerState
+>;
+
+export type BridgeControllerStateChangeEvent = ControllerStateChangeEvent<
+  typeof BRIDGE_CONTROLLER_NAME,
+  BridgeControllerState
+>;
+
 // Maps to BridgeController function names
 export type BridgeControllerActions =
+  | BridgeControllerGetStateAction
   | BridgeControllerAction<BridgeBackgroundAction.SET_CHAIN_INTERVAL_LENGTH>
   | BridgeControllerAction<BridgeBackgroundAction.RESET_STATE>
   | BridgeControllerAction<BridgeBackgroundAction.GET_BRIDGE_ERC20_ALLOWANCE>
@@ -336,10 +348,7 @@ export type BridgeControllerActions =
   | BridgeControllerAction<BridgeBackgroundAction.FETCH_QUOTES>
   | BridgeControllerAction<BridgeUserAction.UPDATE_QUOTE_PARAMS>;
 
-export type BridgeControllerEvents = ControllerStateChangeEvent<
-  typeof BRIDGE_CONTROLLER_NAME,
-  BridgeControllerState
->;
+export type BridgeControllerEvents = BridgeControllerStateChangeEvent;
 
 export type AllowedActions =
   | AccountsControllerGetSelectedMultichainAccountAction
@@ -356,10 +365,8 @@ export type AllowedEvents = never;
 /**
  * The messenger for the BridgeController.
  */
-export type BridgeControllerMessenger = RestrictedMessenger<
+export type BridgeControllerMessenger = Messenger<
   typeof BRIDGE_CONTROLLER_NAME,
   BridgeControllerActions | AllowedActions,
-  BridgeControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  BridgeControllerEvents | AllowedEvents
 >;